fix up circular imports, test from_json of various classes
