+++ notion_client/lib/__init__.py
@@ -1,0 +1,2 @@
+from .api_objects import Page, User
+from .database import Database
